Create Makaan Real Estate Properties in India
This is the final dashboard created from a live real estate dataset taken from Makaan.com made available in Kaggle. The dataset was cleaned in excel and inactive and non-verified listings were removed. The project was also verified for data  analysis using Microsoft SQL Server Management Studio. Check the SQL for analysis and check the tableau for dashboard which is the end product of this data analysis project.